add new letsencrypt root cert and libraries
The LetsEncrypt root certificate was not pushed for some reason in the
last release, it was only in the local project and not added to the
repository. For this reason the plugin was not working correctly on a
new machine when the source files were pulled from the repository and
compiled+run. Since letsencrypt just updated their root certificates, I
just added the new root certificate to the project, and it will be valid
until 2040. Also, the necessary libraries were not pushed to the
repository, so now adding and pushing those also.