refactor and add new Preferences vars
even if this commit is partial, I need to push so that I can continue working from my home computer during the weekend. Note that OptionsFrame.java is still in the works, especially the javascript portion of the preview document, I'm still formatting it (borrowed from the Word add-in vb.net script). I also staretd creating LocalizedBibleBooks.java but it is also incomplete!